Move btf constant fetcher to usual fetchers
